+++ dockerfile.gradual
@@ -14,4 +14,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 EXPOSE 80
 
 # Run app.py when the container launches
-CMD ["python", "./ContainerMains/main_gradual.py"]
+CMD ["python", "./src/main_gradual.py"]
