+++ dockerfile.sudden
@@ -14,4 +14,4 @@ RUN pip install --no-cache-dir -r requirements.txt
 EXPOSE 80
 
 # Run app.py when the container launches
-CMD ["python", "./ContainerMains/main_sudden.py"]
+CMD ["python", "./scr/main_sudden.py"]
